+++ src/main/java/insane96mcp/enhancedai/modules/illager/shoot/EAPillagerAttackGoal.java
@@ -35,11 +35,16 @@ public class EAPillagerAttackGoal extends Goal {
     private int attackDelay;
     private int updatePathDelay;
 
-    public EAPillagerAttackGoal(Pillager pMob, double pSpeedModifier, float pAttackRadius) {
+    private int attackCooldown;
+    private float inaccuracy;
+
+    public EAPillagerAttackGoal(Pillager pMob, double pSpeedModifier, float pAttackRadius, int attackCooldown, float inaccuracy) {
         this.crossbowState = CrossbowState.UNCHARGED;
         this.mob = pMob;
         this.speedModifier = pSpeedModifier;
         this.attackRadiusSqr = pAttackRadius * pAttackRadius;
+        this.attackCooldown = attackCooldown;
+        this.inaccuracy = inaccuracy;
         this.setFlags(EnumSet.of(Flag.MOVE, Flag.LOOK));
     }
 
@@ -69,7 +74,6 @@ public class EAPillagerAttackGoal extends Goal {
             this.mob.setChargingCrossbow(false);
             CrossbowItem.setCharged(this.mob.getUseItem(), false);
         }
-
     }
 
     public boolean requiresUpdateEveryTick() {
@@ -121,7 +125,7 @@ public class EAPillagerAttackGoal extends Goal {
             if (useTicks >= CrossbowItem.getChargeDuration(crossbow)) {
                 this.mob.releaseUsingItem();
                 this.crossbowState = CrossbowState.CHARGED;
-                this.attackDelay = 20 + this.mob.getRandom().nextInt(20);
+                this.attackDelay = this.attackCooldown;
                 this.mob.setChargingCrossbow(false);
             }
         }
@@ -132,7 +136,7 @@ public class EAPillagerAttackGoal extends Goal {
             }
         }
         else if (this.crossbowState == CrossbowState.READY_TO_ATTACK && hasLineOfSight) {
-            this.performCrossbowAttack(1.6F /*TODO Configurable?*/);
+            this.performCrossbowAttack();
             ItemStack crossbow = this.mob.getItemInHand(ProjectileUtil.getWeaponHoldingHand(this.mob, (item) -> item instanceof CrossbowItem));
             CrossbowItem.setCharged(crossbow, false);
             this.crossbowState = CrossbowState.UNCHARGED;
@@ -144,17 +148,17 @@ public class EAPillagerAttackGoal extends Goal {
         return this.crossbowState == CrossbowState.UNCHARGED;
     }
 
-    private void performCrossbowAttack(float velocity) {
+    private void performCrossbowAttack() {
         InteractionHand interactionhand = ProjectileUtil.getWeaponHoldingHand(this.mob, item -> item instanceof CrossbowItem);
         ItemStack itemstack = this.mob.getItemInHand(interactionhand);
         if (this.mob.isHolding(is -> is.getItem() instanceof CrossbowItem)) {
-            performShooting(this.mob.level(), this.mob, interactionhand, itemstack, velocity, (float) (14 - this.mob.level().getDifficulty().getId() * 4));
+            performShooting(this.mob.level(), this.mob, interactionhand, itemstack, this.inaccuracy);
         }
 
         this.mob.onCrossbowAttackPerformed();
     }
 
-    public static void performShooting(Level pLevel, LivingEntity pShooter, InteractionHand pUsedHand, ItemStack crossbowStack, float pVelocity, float pInaccuracy) {
+    public static void performShooting(Level pLevel, LivingEntity pShooter, InteractionHand pUsedHand, ItemStack crossbowStack, float inaccuracy) {
         if (pShooter instanceof Player player && net.minecraftforge.event.ForgeEventFactory.onArrowLoose(crossbowStack, pShooter.level(), player, 1, true) < 0)
             return;
         List<ItemStack> list = CrossbowItem.getChargedProjectiles(crossbowStack);
@@ -165,13 +169,13 @@ public class EAPillagerAttackGoal extends Goal {
             boolean flag = pShooter instanceof Player && ((Player) pShooter).getAbilities().instabuild;
             if (!itemstack.isEmpty()) {
                 if (i == 0) {
-                    shootProjectile(pLevel, pShooter, pUsedHand, crossbowStack, itemstack, afloat[i], flag, pVelocity, pInaccuracy, 0.0F);
+                    shootProjectile(pLevel, pShooter, pUsedHand, crossbowStack, itemstack, afloat[i], flag, inaccuracy, 0.0F);
                 }
                 else if (i == 1) {
-                    shootProjectile(pLevel, pShooter, pUsedHand, crossbowStack, itemstack, afloat[i], flag, pVelocity, pInaccuracy, -10.0F);
+                    shootProjectile(pLevel, pShooter, pUsedHand, crossbowStack, itemstack, afloat[i], flag, inaccuracy, -10.0F);
                 }
                 else if (i == 2) {
-                    shootProjectile(pLevel, pShooter, pUsedHand, crossbowStack, itemstack, afloat[i], flag, pVelocity, pInaccuracy, 10.0F);
+                    shootProjectile(pLevel, pShooter, pUsedHand, crossbowStack, itemstack, afloat[i], flag, inaccuracy, 10.0F);
                 }
             }
         }
@@ -189,7 +193,7 @@ public class EAPillagerAttackGoal extends Goal {
         return 1.0F / (pRandom.nextFloat() * 0.5F + 1.8F) + f;
     }
 
-    private static void shootProjectile(Level pLevel, LivingEntity pShooter, InteractionHand pHand, ItemStack pCrossbowStack, ItemStack pAmmoStack, float pSoundPitch, boolean pIsCreativeMode, float pVelocity, float pInaccuracy, float pProjectileAngle) {
+    private static void shootProjectile(Level pLevel, LivingEntity pShooter, InteractionHand pHand, ItemStack pCrossbowStack, ItemStack pAmmoStack, float pSoundPitch, boolean pIsCreativeMode, float inaccuracy, float pProjectileAngle) {
         if (!pLevel.isClientSide) {
             boolean isShootingFirework = pAmmoStack.is(Items.FIREWORK_ROCKET);
             Projectile projectile;
@@ -204,7 +208,7 @@ public class EAPillagerAttackGoal extends Goal {
             }
 
             if (pShooter instanceof CrossbowAttackMob mob) {
-                attackEntityWithRangedAttack(pShooter, mob.getTarget(), pCrossbowStack, projectile, pProjectileAngle);
+                attackEntityWithRangedAttack(pShooter, mob.getTarget(), pCrossbowStack, projectile, pProjectileAngle, inaccuracy);
             }
 
             pCrossbowStack.hurtAndBreak(isShootingFirework ? 3 : 1, pShooter, entity -> entity.broadcastBreakEvent(pHand));
@@ -213,22 +217,19 @@ public class EAPillagerAttackGoal extends Goal {
         }
     }
 
-    protected static void attackEntityWithRangedAttack(LivingEntity livingEntity, LivingEntity target, ItemStack crossbow, Projectile projectile, float angle) {
+    protected static void attackEntityWithRangedAttack(LivingEntity livingEntity, LivingEntity target, ItemStack crossbow, Projectile projectile, float angle, float inaccuracy) {
         double distance = livingEntity.distanceTo(target);
-        double distanceY = (float) (target.getY() - livingEntity.getY());
+        float distanceY = (float) (target.getY() - livingEntity.getY());
         float dirX = (float) (target.getX() - livingEntity.getX());
         float dirZ = (float) (target.getZ() - livingEntity.getZ());
         float distanceXZ = (float) Math.sqrt(dirX * dirX + dirZ * dirZ);
         float yPos = (float) target.getY(0d);
-        yPos += target.getEyeHeight() * 0.5;
+        yPos += target.getEyeHeight() * 0.5f;
         if (distanceXZ != 0f)
-            yPos += (distanceY / distanceXZ);
+            yPos += distanceY / distanceXZ;
         float dirY = (float) (yPos - projectile.getY());
         Vector3f shootRotation = ((CrossbowAttackMob) livingEntity).getProjectileShotVector(livingEntity, new Vec3(dirX, dirY + distanceXZ * 0.19f, dirZ), angle);
-        //Vector3f shootRotation = new Vector3f(dirX, dirY + distanceXZ * 0.19f, dirZ);
-        //shootRotation = (new Vector3f(shootRotation)).rotateAxis(angle * ((float)Math.PI / 180F), shootRotation.x, shootRotation.y, shootRotation.z);
-        projectile.shoot(shootRotation.x(), shootRotation.y(), shootRotation.z(), 1.1f + ((float) distance / 32f) + (float) Math.max(distanceY / 48d, 0f), 0f);
-        livingEntity.playSound(SoundEvents.CROSSBOW_SHOOT, 1.0F, 1.0F / (livingEntity.getRandom().nextFloat() * 0.4F + 0.8F));
+        projectile.shoot(shootRotation.x(), shootRotation.y(), shootRotation.z(), 1.1f + ((float) distance / 32f) + (float) Math.max(distanceY / 48d, 0f), inaccuracy);
     }
 
     enum CrossbowState {