+++ changelog.md
@@ -3,6 +3,7 @@
 ## Upcoming
 * Added Pillager shoot
   * Pillagers can now shoot from farther away
+  * Inaccuracy and shooting cooldown is now configurable
 * Skeleton inaccuracy is now configurable per difficulty
   * Slightly decreased accuracy in hard
 * Fixed Witches throwing potions from farther than their throw range